+++ zellij-server/src/lib.rs
@@ -17,6 +17,7 @@ use std::{
     sync::{Arc, Mutex, RwLock},
     thread,
 };
+use zellij_utils::nix::sys::stat::{umask, Mode};
 use zellij_utils::pane_size::Size;
 use zellij_utils::zellij_tile;
 
@@ -65,6 +66,7 @@ pub(crate) enum ServerInstruction {
     ClientExit(ClientId),
     RemoveClient(ClientId),
     Error(String),
+    KillSession,
     DetachSession(ClientId),
     AttachClient(ClientAttributes, Options, ClientId),
 }
@@ -78,6 +80,7 @@ impl From<&ServerInstruction> for ServerContext {
             ServerInstruction::ClientExit(..) => ServerContext::ClientExit,
             ServerInstruction::RemoveClient(..) => ServerContext::RemoveClient,
             ServerInstruction::Error(_) => ServerContext::Error,
+            ServerInstruction::KillSession => ServerContext::KillSession,
             ServerInstruction::DetachSession(..) => ServerContext::DetachSession,
             ServerInstruction::AttachClient(..) => ServerContext::AttachClient,
         }
@@ -173,9 +176,13 @@ impl SessionState {
 
 pub fn start_server(mut os_input: Box<dyn ServerOsApi>, socket_path: PathBuf) {
     info!("Starting Zellij server!");
+
+    // preserve the current umask: read current value by setting to another mode, and then restoring it
+    let current_umask = umask(Mode::all());
+    umask(current_umask);
     daemonize::Daemonize::new()
         .working_directory(std::env::current_dir().unwrap())
-        .umask(0o077)
+        .umask(current_umask.bits())
         .start()
         .expect("could not daemonize the server process");
 
@@ -399,6 +406,14 @@ pub fn start_server(mut os_input: Box<dyn ServerOsApi>, socket_path: PathBuf) {
                         .send_to_screen(ScreenInstruction::RemoveClient(client_id))
                         .unwrap();
                 }
+            }
+            ServerInstruction::KillSession => {
+                let client_ids = session_state.read().unwrap().client_ids();
+                for client_id in client_ids {
+                    os_input.send_to_client(client_id, ServerToClientMsg::Exit(ExitReason::Normal));
+                    remove_client!(client_id, os_input, session_state);
+                }
+                break;
             }
             ServerInstruction::DetachSession(client_id) => {
                 os_input.send_to_client(client_id, ServerToClientMsg::Exit(ExitReason::Normal));